+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-11-09
+- cancelOrder added reason parameter
+- auto-create 'data' folder if does not exist
+
 ## [1.1.0] - 2022-07-17
 - setLogger to inject logger
 - updated dependencies: better-sqlite3, axios, sockjs-client, keytar